backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.50
python-dotenv==1.0.1
sse-starlette==2.2.1
httpx==0.27.2
pydantic==2.9.2
